Check Delete Function

+++ app/controllers/question/index.controller.js
@@ -124,15 +124,23 @@ class index {
     delete_question(req, res) {
         var r = req.r;
         var params = req.query;
-
-        r.db('lms').table('question').get(params.id).delete()
+        r.db('lms').table('exam_test_detail').filter({ question_id: params.id })
             .run()
-            .then(function (result) {
-                res.json(result);
+            .then(function (re) {
+                if (re.length > 0) {
+                    res.status(500).json({error:"ข้อสอบนี้มีการใช้งาน"});
+                } else {
+                    r.db('lms').table('question').get(params.id).delete()
+                        .run()
+                        .then(function (result) {
+                            res.json(result);
+                        })
+                        .catch(function (err) {
+                            res.status(500).json(err);
+                        })
+                }
             })
-            .catch(function (err) {
-                res.status(500).json(err);
-            })
+
     }
 
     confirm(req, res) {
